requirements.txt
# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0

# Text Processing and NLP
nltk>=3.8.0
spacy>=3.6.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
flake8>=6.0.0
black>=23.0.0

# Data visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0